minor pre-user data functionality.  Need to figure out use cases and paths to edit / add data.  The use cases for editing a set (1) new, (2) may data, and (3) the graph (prov panel via "view raw data")
Note:  The problem with editing from a Graph's quickView is that the quickView plot may be a composite.  What to edit then, the component, the plot, the component set, or the plot set?  That would be an advanced capability to come in future Workbench versions.

Version 1.0 solution is to turn the "view source data" into "edit data".  This is more